[EJBCLIENT-120] Hang in AbstractHandleableCloseable.close() when using @Clustered Extract the logic from cluster context outside EJBClientContext for uncoupling the operations between EJBClientContext and ClusterContext. This removes the deadlock when both types of operations are being invoked through the EJBClientContext.
